requirements.txt
litellm 
python-dotenv 